Rename built-in subscriber file
