+++ dds/src/implementation/dds_impl/mod.rs
@@ -6,7 +6,7 @@ pub mod builtin_stateful_reader;
 pub mod builtin_stateful_writer;
 pub mod builtin_stateless_reader;
 pub mod builtin_stateless_writer;
-pub mod builtin_subscriber;
+pub mod builtin_subscriber_impl;
 pub mod dcps_service;
 pub mod domain_participant_impl;
 pub mod message_receiver;
